Add plugin priority and refactor for compatibility with napari-ndev legacy behavior (#24)
# References and relevant issues

This PR is needed to maintain behavior for using `ndevio` in
`napari-ndev` in https://github.com/ndev-kit/napari-ndev/pull/194

# Description

1. Adds `bioio-tifffile` to core dependencies, because, at least for
now, this is needed for some awkward tiff files that _can_ be read by
`bioio-imageio` but result in bad metadata reading. Since there are
currently no `tifffile` pins in any of the dependencies, this behavior
will remain for now. It's worth considering in the future if this is
really the way I want core dependencies going forward.
2. This orders the plugin list to be used for infering the reader,
because despite the installation order of plugins, some get installed
before others (perhaps this is a `uv` thing, compared to `pip`). As
such, I need a way to opinionatedly pick my preferred plugins, but the
preferred plugin setting still gives the user an easy way to override if
they can only find their desired behavior with a specific plugin. It
will always fall back to the setting.
3. Refactors `determine_reader_plugin` to mostly just rely on the plugin
manager